Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 
 int main() {
-    int a, b;
+    float a, b;
 
 
-    scanf("%d %d", &a, &b);
+    scanf("%f %f", &a, &b);
 
-    printf("Product: %d\n", a * b);
+    printf("Product: %.2f\n", a * b);
 
     return 0;
 }
